Move MESSAGE_PERSONS_LISTED_OVERVIEW to FindCommand
FindCommand is the only command that uses this message.

Usually, messages that a specific to a command are kept in the command
itself. For example, MESSAGE_DELETE_PERSON_SUCCESS is a member of
DeleteCommand, since it is the only user of that message.

As such, to be consistent, let's move MESSAGE_PERSONS_LISTED_OVERVIEW to
FindCommand.
